// src/api/types/MyPatientResponse.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

export type MyPatientResponse = {
  /**
   * @type integer | undefined, int64
   */
  userId?: number
  /**
   * @type string | undefined
   */
  phoneNumber?: string
  /**
   * @type string | undefined
   */
  fullName?: string
  /**
   * @type string | undefined
   */
  gender?: string
  /**
   * @type string | undefined, date
   */
  dateOfBirth?: string
  /**
   * @type string | undefined, date-time
   */
  assignedAt?: string
  /**
   * @type integer | undefined, int32
   */
  assignedCoursesCount?: number
  /**
   * @type integer | undefined, int32
   */
  completedCoursesCount?: number
  /**
   * @type boolean | undefined
   */
  hasActiveCourse?: boolean
}